third_party/WebKit/LayoutTests/http/tests/htmlimports/import-dynamic-descendant-block.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
<link id="parentOfDynamicChild" rel="import" href="resources/having-dynamic-child.html">
<script>
test(function() {
   assert_true(null != parentOfDynamicChild.import, 'Parent of dynamic child'); 
   assert_true(null != parentOfDynamicChild.import.querySelector('#dynamicLink').import, 'Dynamically added child'); 
}, 'Dynamically added import in a descendant blocks.');
</script>
</body>
</html>
